add: some tests in bst and add methods
