+++ homework-testing/src/binary-search-tree.js
@@ -1,4 +1,4 @@
-import Node from './node';
+const Node = require('./node');
 
 class BinarySearchTree {
   constructor() {
@@ -21,6 +21,38 @@ class BinarySearchTree {
     }
   }
 
+  delete(key) {
+    if (!Number.isInteger(key)) {
+      return;
+    }
+    this.root = this.deleteNode(this.root, key);
+  }
+
+  search(key) {
+    if (!Number.isInteger(key)) {
+      return false;
+    }
+    const result = this.searchHelper(this.root, key);
+    return result;
+  }
+
+  contains(value) {
+    const result = this.fromLeftToRightTraverse(this.root);
+    return result.includes(value);
+  }
+
+  traverse(order) {
+    if (order) {
+      return this.fromLeftToRightTraverse(this.root);
+    }
+    return this.fromRightToLeftTraverse(this.root);
+  }
+
+  verify() {
+    const result = this.verifyHelper(this.root);
+    return result;
+  }
+
   insertNode(node, newNode) {
     if (newNode.key < node.key) {
       if (node.left === null) {
@@ -39,15 +71,10 @@ class BinarySearchTree {
     }
   }
 
-  delete(key) {
-    if (!Number.isInteger(key)) {
-      return;
-    }
-    this.root = this.deleteNode(this.root, key);
-  }
-
   deleteNode(node, key) {
-    if (node === null) return null;
+    if (node === null) {
+      return null;
+    }
     if (key < node.key) {
       node.left = this.deleteNode(node.left, key);
       return node;
@@ -60,7 +87,6 @@ class BinarySearchTree {
       node = null;
       return node;
     }
-
     if (node.left === null) {
       node = node.right;
       return node;
@@ -69,62 +95,87 @@ class BinarySearchTree {
       node = node.left;
       return node;
     }
-
     const aux = this.findMinimumNode(node.right);
     node.key = aux.key;
-
     node.right = this.deleteNode(node.right, aux.key);
     return node;
   }
 
-  find(node, key) {
+  searchHelper(node, key) {
     if (node === null) {
       return null;
     }
-    if (key < node.key) {
-      return this.find(node.left, key);
-    }
-    if (key > node.key) {
-      return this.find(node.right, key);
-    }
-    return node;
-  }
-
-  contains(key) {
-    let current = this.root;
-    while (current) {
-      console.log(current);
-      if (key === current.key) {
-        return true;
-      }
-      if (key < current.key) {
-        current = current.left;
-      }
-      if (key > current.key) {
-        current = current.right;
-      }
-    }
-    return false;
-  }
-
-  search(key) {
-    if (!Number.isInteger(key)) {
-    } else {
-      return this.searchHelper(this.root, key);
-    }
-  }
-
-  searchHelper(node, key) {
     if (node.key === key) {
       return node;
     }
-    if (key < node.key && node.left !== null) {
-      console.log(node.left);
-      this.searchHelper(node.left, key);
-    } else if (key > node.key && node.right !== null) {
-      this.searchHelper(node.right, key);
+    if (key < node.key && key !== node.key) {
+      if (node.left !== null) {
+        return this.searchHelper(node.left, key);
+      }
+    } else if (key > node.key && key !== node.key) {
+      if (node.right !== null) {
+        return this.searchHelper(node.right, key);
+      }
     }
+  }
+
+  fromLeftToRightTraverse(traverseNode) {
+    const output = [];
+
+    function inOrder(node) {
+      if (node === null) {
+        return;
+      }
+      inOrder(node.left);
+      output.push(node.value);
+      inOrder(node.right);
+    }
+    inOrder(traverseNode);
+
+    return output;
+  }
+
+  fromRightToLeftTraverse(traverseNode) {
+    const output = [];
+
+    function notInOrder(node) {
+      if (node === null) {
+        return;
+      }
+      notInOrder(node.right);
+      output.push(node.value);
+      notInOrder(node.left);
+    }
+    notInOrder(traverseNode);
+
+    return output;
+  }
+
+  findMinimumNode(node) {
+    if (node.left === null) {
+      return node;
+    }
+    return this.findMinimumNode(node.left);
+  }
+
+  verifyHelper(node, min = null, max = null) {
+    if (node === null) {
+      return null;
+    }
+    if (max !== null && node.key > max) {
+      return false;
+    }
+    if (min !== null && node.key < min) {
+      return false;
+    }
+    if (node.left && !this.verifyHelper(node.left, min, node.key)) {
+      return false;
+    }
+    if (node.tight && !this.verifyHelper(node.right, node.key, max)) {
+      return false;
+    }
+    return true;
   }
 }
 
-export default BinarySearchTree;
+module.exports = BinarySearchTree;
