+++ homework-testing/src/node.js
@@ -7,4 +7,4 @@ class Node {
   }
 }
 
-export default Node;
+module.exports = Node;
